+++ src/cpu/vector_engine/vpu/multilane_wrapper/func_unit.hh
@@ -31,7 +31,7 @@
 #include "debug/Datapath.hh"
 
 float
-Datapath::compute_float_fp_op(float Aitem, float Bitem, int Mitem,
+Datapath::compute_float_fp_op(float Aitem, float Bitem, uint8_t Mitem,
     float Dstitem,  RiscvISA::VectorStaticInst* insn)
 {
     float Ditem=0;
@@ -146,7 +146,7 @@ Datapath::compute_float_fp_op(float Aitem, float Bitem, int Mitem,
 
 double
 Datapath::compute_double_fp_op(double Aitem, double Bitem,
-     long int Mitem, double Dstitem,  RiscvISA::VectorStaticInst* insn)
+     uint8_t Mitem, double Dstitem,  RiscvISA::VectorStaticInst* insn)
 {
     float Ditem=0;
     std::string operation = insn->getName();
@@ -258,7 +258,7 @@ Datapath::compute_double_fp_op(double Aitem, double Bitem,
 }
 
 double 
-Datapath::computeDoubleFPReduction(double accumDp,double Bitem,long int Mitem)
+Datapath::computeDoubleFPReduction(double accumDp,double Bitem,uint8_t Mitem)
 {
     double reduction;
     std::string operation = insn->getName();
@@ -284,7 +284,7 @@ Datapath::computeDoubleFPReduction(double accumDp,double Bitem,long int Mitem)
 }
 
 float
-Datapath::computeSingleFPReduction(float accumSp,float Bitem,int Mitem)
+Datapath::computeSingleFPReduction(float accumSp,float Bitem,uint8_t Mitem)
 {
     float reduction;
     std::string operation = insn->getName();
@@ -404,7 +404,7 @@ Datapath::compute_double_fp_comp_op(double Aitem, double Bitem,
 
 long int
 Datapath::compute_long_int_op(long int Aitem, long int Bitem,
-    long int Mitem, long int Dstitem,  RiscvISA::VectorStaticInst* insn)
+    uint8_t Mitem, long int Dstitem,  RiscvISA::VectorStaticInst* insn)
 {
     long int Ditem=0;
     std::string operation = insn->getName();
@@ -567,7 +567,7 @@ Datapath::compute_long_int_op(long int Aitem, long int Bitem,
 }
 
 int
-Datapath::compute_int_op(int Aitem, int Bitem, int Mitem,
+Datapath::compute_int_op(int Aitem, int Bitem, uint8_t Mitem,
         int Dstitem,  RiscvISA::VectorStaticInst* insn)
 {
     int Ditem=0;
@@ -856,7 +856,7 @@ Datapath::computeIntMaskLogicalOp(bool Aitem, bool Bitem,
 }
 
 double
-Datapath::compute_cvt_f_x_64_op( long int Bitem, long int Mitem,
+Datapath::compute_cvt_f_x_64_op( long int Bitem, uint8_t Mitem,
     long int Dstitem,  RiscvISA::VectorStaticInst* insn)
 {
     double Ditem=0;
@@ -878,7 +878,7 @@ Datapath::compute_cvt_f_x_64_op( long int Bitem, long int Mitem,
 }
 
 float
-Datapath::compute_cvt_f_x_32_op( int Bitem, int Mitem, int Dstitem,
+Datapath::compute_cvt_f_x_32_op( int Bitem, uint8_t Mitem, int Dstitem,
     RiscvISA::VectorStaticInst* insn)
 {
     float Ditem=0;
@@ -900,7 +900,7 @@ Datapath::compute_cvt_f_x_32_op( int Bitem, int Mitem, int Dstitem,
 }
 
 long int
-Datapath::compute_cvt_x_f_64_op( double Bitem, long int Mitem,
+Datapath::compute_cvt_x_f_64_op( double Bitem, uint8_t Mitem,
     double Dstitem, RiscvISA::VectorStaticInst* insn)
 {
     long int Ditem=0;
@@ -922,7 +922,7 @@ Datapath::compute_cvt_x_f_64_op( double Bitem, long int Mitem,
 }
 
 int
-Datapath::compute_cvt_x_f_32_op( float Bitem, int Mitem,
+Datapath::compute_cvt_x_f_32_op( float Bitem, uint8_t Mitem,
     float Dstitem, RiscvISA::VectorStaticInst* insn)
 {
     int Ditem=0;
